requirements.txt
# LexAI Practice Partner - Production Dependencies

# Core Flask framework
Flask==3.1.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.0

# Database migrations
Flask-Migrate>=4.0.7
Flask-Script>=2.0.6

# Security and authentication
bcrypt>=4.3.0
Flask-WTF>=1.2.1
WTForms>=3.1.1

# Database drivers
psycopg2-binary>=2.9.3  # PostgreSQL
sqlite3  # SQLite (built-in)

# Environment and configuration
python-dotenv>=1.0.1

# HTTP requests and API integration
requests>=2.32.3
urllib3>=2.4.0

# File processing
PyPDF2>=3.0.1
python-magic>=0.4.27  # File type detection

# Rate limiting and caching
Flask-Limiter>=3.8.0
redis>=5.1.1

# Email support
Flask-Mail>=0.9.1

# Production server
gunicorn>=22.0.0
gevent>=24.2.1

# Monitoring and logging
sentry-sdk[flask]>=2.15.0

# Development dependencies (optional)
pytest>=8.3.3
pytest-flask>=1.3.0
pytest-cov>=5.0.0
black>=24.0.0
flake8>=7.1.1

# Security scanning
safety>=3.2.8
bandit>=1.7.10